Add styling

+++ public/js/delete-post.js
@@ -1,18 +1,19 @@
 const deletePostFormHandler = async (event) => {
     event.preventDefault();
-  
-    const post_id = document.querySelector('input[name="post-id"]').value;
-  
-    const response = await fetch(`/api/posts/${post_id}`, {
-    method: 'DELETE',
-    body: JSON.stringify({ post_id }),
-    headers: { 'Content-Type': 'application/json' },
-    });
+    if (confirm("Are you sure you want to delete this post?")) {
+        const post_id = document.querySelector('input[name="post-id"]').value;
+        
+        const response = await fetch(`/api/posts/${post_id}`, {
+        method: 'DELETE',
+        body: JSON.stringify({ post_id }),
+        headers: { 'Content-Type': 'application/json' },
+        });
 
-    if (response.ok) {
-        document.location.replace('/dashboard');
-    } else {
-        alert('An error occured.');
+        if (response.ok) {
+            document.location.replace('/dashboard');
+        } else {
+            alert('An error occured.');
+        }
     }
   };
